mod(feat/repository): basic page with repositories
