+++ src/app/repository/page.css
@@ -1,0 +1,31 @@
+/* RepositoryPage.css */
+.container {
+    max-width: 800px;
+    margin: 0 auto;
+    padding: 20px;
+    font-family: Arial, sans-serif;
+}
+
+.title {
+    text-align: center;
+    font-size: 2em;
+    margin-bottom: 20px;
+}
+
+.repository-list {
+    list-style-type: none;
+    padding: 0;
+}
+
+.repository-item {
+    background-color: #f9f9f9;
+    margin: 10px 0;
+    padding: 15px;
+    border-radius: 5px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    transition: background-color 0.3s;
+}
+
+.repository-item:hover {
+    background-color: #e9e9e9;
+}
